Now checking if the workdir exists before looping over it

+++ src/tcutility/job/ams.py
@@ -180,10 +180,17 @@ class AMSJob(Job):
         '''
         os.makedirs(self.rundir, exist_ok=True)
 
-        for file in os.listdir(self.rundir):
-            p = os.path.join(self.rundir, file)
-            if p.endswith('.rkf'):
-                os.remove(p)
+        if os.path.exists(self.workdir):
+            for file in os.listdir(self.workdir):
+                p = os.path.join(self.workdir, file)
+                if p.endswith('.rkf'):
+                    os.remove(p)
+                if 'ams.kid' in p:
+                    os.remove(p)
+                if p.startswith('t21.'):
+                    os.remove(p)
+                if p.startswith('t12.'):
+                    os.remove(p)
 
         if not self._molecule and not self._molecule_path and 'atoms' not in self.settings.input.ams.system:
             log.error(f'You did not supply a molecule for this job. Call the {self.__class__.__name__}.molecule method to add one.')
